requirements.txt
pyyaml
psutil
colorama